Refactor CodeOwner feature + Skip Test with Zero duration.
