+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Testing/Common.cs
@@ -4,8 +4,12 @@
 // </copyright>
 
 using System;
+using System.Reflection;
+using Datadog.Trace.Ci;
+using Datadog.Trace.Ci.Tags;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.Logging;
+using Datadog.Trace.PDBs;
 
 namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing
 {
@@ -20,7 +24,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing
                 return;
             }
 
-            Ci.CIVisibility.FlushSpans();
+            CIVisibility.FlushSpans();
         }
 
         internal static string GetParametersValueData(object paramValue)
@@ -29,7 +33,8 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing
             {
                 return "(null)";
             }
-            else if (paramValue is Array pValueArray)
+
+            if (paramValue is Array pValueArray)
             {
                 const int maxArrayLength = 50;
                 int length = pValueArray.Length > maxArrayLength ? maxArrayLength : pValueArray.Length;
@@ -42,12 +47,32 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing
 
                 return "[" + string.Join(", ", strValueArray) + (pValueArray.Length > maxArrayLength ? ", ..." : string.Empty) + "]";
             }
-            else if (paramValue is Delegate pValueDelegate)
+
+            if (paramValue is Delegate pValueDelegate)
             {
                 return $"{paramValue}[{pValueDelegate.Target}|{pValueDelegate.Method}]";
             }
 
             return paramValue.ToString();
+        }
+
+        internal static void DecorateSpanWithSourceAndCodeOwners(Span span, MethodInfo testMethod)
+        {
+            if (MethodSymbolResolver.Instance.TryGetMethodSymbol(testMethod, out var methodSymbol))
+            {
+                span.SetTag(TestTags.SourceFile, CIEnvironmentValues.Instance.MakeRelativePathFromSourceRoot(methodSymbol.File));
+                span.SetMetric(TestTags.SourceStart, methodSymbol.StartLine);
+                span.SetMetric(TestTags.SourceEnd, methodSymbol.EndLine);
+
+                if (CIEnvironmentValues.Instance.CodeOwners is { } codeOwners)
+                {
+                    var match = codeOwners.Match("/" + CIEnvironmentValues.Instance.MakeRelativePathFromSourceRoot(methodSymbol.File, false));
+                    if (match is not null)
+                    {
+                        span.SetTag(TestTags.CodeOwners, match.Value.GetOwnersString());
+                    }
+                }
+            }
         }
     }
 }
